refactor: 공통으로 사용하는 요소들 따로 common 패키지로 분리
