+++ src/main/java/subway/common/exception/NotSupportedFunctionException.java
@@ -1,4 +1,4 @@
-package subway.station.exception;
+package subway.common.exception;
 
 public class NotSupportedFunctionException extends IllegalArgumentException {
     private static final String MESSAGE = "선택할 수 없는 기능입니다.";
