+++ src/main/java/subway/common/util/InputTypeValidator.java
@@ -1,4 +1,4 @@
-package subway.util;
+package subway.common.util;
 
 import java.util.regex.Pattern;
