+++ src/main/java/subway/station/StationFunctionMapper.java
@@ -1,6 +1,6 @@
 package subway.station;
 
-import subway.station.exception.NotSupportedFunctionException;
+import subway.common.exception.NotSupportedFunctionException;
 
 import java.util.HashMap;
 import java.util.Map;
